Update README.md
I added common build problems to the common issues section of readme in an attempt to decrease the number of issues incoming from a user base that is continuously getting larger.
Also I aded a note to ping me for build questions because otherwise questions in the obs discord probably wouldn't be seen. I just see the problem here that I'm not always available (duh). Maybe we should ask for a "mac-virtualcam-support role" on the discord that could be pinged and whoever wants to get spammed on discord get's it? 😄